Adding analysis scripts
Plots duration, numbers, latency, responses in the same plot. Colours
are similar for the same type of cue, darker are from rewarded trials.

Base functions (eg. Epoch methods) are mostly tested, but not necessarily
with this analysis in mind. Need to add more tests to those before this
analysis can be trusted. Also loading MedPC scripts seem to have expected
behavior, but no formal tests are included to those yet.