Handles broken session 1
Note: replaces missing trials with mean values for that trial
type in that session. It's a hack, but shouldn't affect things too
much. If needed, we could return the unaltered df with the nans for
missing trials as well as this one and only use this for plotting
while the other can be used for stats