Adding rats (RH05 and R105) for analysis
R105 had issues during the experiment such that he did not receive
the correct (balanced) number of steady lights and flashing lights.
In the experimental control, trial 3 had the flashing light on timer
turn on followed by the steady light off timer. Can parse this data
if necessary, but probably easiest to start from the next session.

RH05 has an unbalanced number of photobeam on/off breaks. Can be fixed
by finding the off double and removing it.