Cue epochs based on cue duration instead of on/off times
The Neuralynx on/off postevent times was not always posted. Unsure
under what situation resulted in the event not posted (will try to
reset MATLAB between experiments, but the missing Off time wasn't
the first of that type). Here, I'm just specifying the cue_duration
(10 sec), and determining the end of the epoch using that from the
On time. This will be more accurate stats than using the on/off
times (they're current up to 100 ms off - for the flashing light set
to 0.1 pulse duration, I think there might be an interaction between
the timers that limit pin changes in Neuralynx). The only issue would
be if an On event is missing, currently that would go unnoticed.