Restructure code for new experiment
This includes some code modifications as well as shuffling so
that it's more obvious which script is doing which analysis.
During this training session, 7 of the rats are being trained in
the medpc boxes in the bucci room, 1 rat is being trained in the
recording box in the vdmlab room.

- Individual rat labels (instead of 1 - 8).

- Specified group1/group2 as input to Rat class.

- Neuralynx event labels given to vdm.load_events as a dict.

- Renamed variables to be more generic (eg. cue becomes light1).

- Add a 'check_experiment' notebook with some basic checks (eg. n_trials).