Organize plot output and single session check
Single session check verifies that there are X number of trials
and cues for the most recent experiment (X as expected for this
protocol).

Plot output was renamed to utilize the RXXX naming system that these
rats follow for this biconditional training. Output is saved to a
'winter2017' folder in the 'plots' folder to keep things organized.